cleant as fuck

+++ test/const.cc
@@ -1,5 +1,5 @@
   int main ()
   {
-    int i = 1;
-    while (i < 10) i++ ;
+   double i = 1.0;
+    while (i < 10.0) i++ ;
   }
